+++ packages/isar/lib/src/native/open.dart
@@ -5,11 +5,12 @@ import 'dart:math';
 import 'package:ffi/ffi.dart';
 import 'package:isar/isar.dart';
 
+import 'bindings.dart';
 import 'isar_collection_impl.dart';
 import 'isar_core.dart';
 import 'isar_impl.dart';
 
-final _isarPtrPtr = malloc<Pointer>();
+final _isarPtrPtr = malloc<Pointer<CIsarInstance>>();
 
 void _initializeInstance(
     Allocator alloc, IsarImpl isar, List<CollectionSchema> schemas) {
@@ -17,7 +18,7 @@ void _initializeInstance(
       .map((e) => e.propertyIds.length)
       .reduce((value, element) => max(value, element));
 
-  final colPtrPtr = alloc<Pointer>();
+  final colPtrPtr = alloc<Pointer<CIsarCollection>>();
   final offsetsPtr = alloc<Uint32>(maxProperties);
 
   final cols = <Type, IsarCollection>{};
@@ -45,18 +46,18 @@ void _initializeInstance(
 
 Future<Isar> openIsar({
   required List<CollectionSchema> schemas,
-  required String directory,
+  String? directory,
   String name = 'isar',
   bool relaxedDurability = true,
 }) async {
-  initializeIsarCore();
-  IC.isar_connect_dart_api(NativeApi.postCObject);
+  initializeCoreBinary();
+  IC.isar_connect_dart_api(NativeApi.postCObject.cast());
 
   return using((alloc) async {
     final namePtr = name.toNativeUtf8(allocator: alloc);
-    final dirPtr = directory.toNativeUtf8(allocator: alloc);
+    final dirPtr = directory?.toNativeUtf8(allocator: alloc) ?? nullptr;
 
-    final schemaStr = '[' + schemas.map((e) => e.schema).join(',') + ']';
+    final schemaStr = '[${schemas.map((e) => e.schema).join(',')}]';
     final schemaStrPtr = schemaStr.toNativeUtf8(allocator: alloc);
 
     final receivePort = ReceivePort();
@@ -74,18 +75,18 @@ Future<Isar> openIsar({
 
 Isar openIsarSync({
   required List<CollectionSchema> schemas,
-  required String directory,
+  String? directory,
   String name = 'isar',
   bool relaxedDurability = true,
 }) {
-  initializeIsarCore();
-  IC.isar_connect_dart_api(NativeApi.postCObject);
+  initializeCoreBinary();
+  IC.isar_connect_dart_api(NativeApi.postCObject.cast());
 
   return using((alloc) {
     final namePtr = name.toNativeUtf8(allocator: alloc);
-    final dirPtr = directory.toNativeUtf8(allocator: alloc);
+    final dirPtr = directory?.toNativeUtf8(allocator: alloc) ?? nullptr;
 
-    final schemaStr = '[' + schemas.map((e) => e.schema).join(',') + ']';
+    final schemaStr = '[${schemas.map((e) => e.schema).join(',')}]';
     final schemaStrPtr = schemaStr.toNativeUtf8(allocator: alloc);
 
     nCall(IC.isar_create_instance(_isarPtrPtr, namePtr.cast(), dirPtr.cast(),