+++ packages/isar/tool/generate_bindings.sh
@@ -2,14 +2,16 @@
 
 script_dir=$(cd "$(dirname "$0")"; pwd -P)
 
-tmp_dir=$(mktemp -d 2>/dev/null || mktemp -d -t 'mytmpdir')
-core_version=`cat ../../CORE_VERSION`
+tmp_dir=$(mktemp -d 2>/dev/null || mktemp -d -t 'tmpdir')
+core_version=`dart $script_dir/get_version.dart`
+
+echo $core_version
 
 cd $tmp_dir
 
 git clone https://github.com/isar/isar-core.git
 cd isar-core
-git checkout $core_version
+git checkout tags/$core_version
 git submodule update --init
 
 cd dart-ffi
@@ -25,5 +27,4 @@ dart pub get
 dart run ffigen --config tool/ffigen.yaml
 rm isar-dart.h
 
-dart tool/fix_bindings.dart
 dart format --fix lib/src/native/bindings.dart