+++ packages/isar/tool/get_version.dart
@@ -1,0 +1,7 @@
+// ignore_for_file: avoid_relative_lib_imports
+
+import '../lib/src/version.dart';
+
+void main() {
+  print(isarCoreVersion);
+}