+++ packages/isar_generator/lib/src/code_gen/collection_schema_generator.dart
@@ -82,6 +82,7 @@ String _generateSchema(ObjectInfo object) {
         {
           'name': index.name,
           'unique': index.unique,
+          'replace': index.replace,
           'properties': [
             for (var indexProperty in index.properties)
               {