+++ packages/isar_generator/lib/src/code_gen/type_adapter_generator_common.dart
@@ -42,5 +42,5 @@ String generateAttachLinks(ObjectInfo object) {
       id
     );''';
   }
-  return code + '}';
+  return '$code}';
 }