+++ packages/isar_test/test/util/common.dart
@@ -37,12 +37,23 @@ Future qEqualSync<T>(List<T> actual, List<T> target) async {
 var allTestsSuccessful = true;
 var testCount = 0;
 
+Future<void> _prepareTest() async {
+  if (!kIsWeb) {
+    try {
+      await Isar.initializeIsarCore(download: true);
+    } catch (e) {
+      // ignore. maybe this is an instrumentation test
+    }
+  }
+}
+
 @isTest
 void isarTest(String name, dynamic Function() body, {Timeout? timeout}) {
   test(
     name,
     () async {
       try {
+        await _prepareTest();
         await body();
         testCount++;
       } catch (e) {
@@ -58,6 +69,7 @@ void isarTest(String name, dynamic Function() body, {Timeout? timeout}) {
 void isarTestVm(String name, dynamic Function() body) {
   test(name, () async {
     try {
+      await _prepareTest();
       await body();
       testCount++;
     } catch (e) {
@@ -67,37 +79,21 @@ void isarTestVm(String name, dynamic Function() body) {
   }, skip: kIsWeb);
 }
 
-String? testTempPath;
-
-// Change the path for your platform to point to a local isar core build
-void registerBinaries() {
-  if (!kIsWeb && testTempPath == null) {
-    final dartToolDir = path.join(Directory.current.path, '.dart_tool');
-    testTempPath = path.join(dartToolDir, 'test', 'tmp');
-    try {
-      Isar.initializeLibraries(
-        libraries: {
-          IsarAbi.windowsX64: path.join(dartToolDir, 'libisar_windows_x64.dll'),
-          IsarAbi.macosX64: path.join(dartToolDir, 'libisar_macos.dylib'),
-          IsarAbi.linuxArm64: path.join(dartToolDir, 'libisar_linux_x64.so'),
-        },
-      );
-    } catch (e) {
-      // ignore. maybe this is an instrumentation test
-    }
-  }
-}
-
 String getRandomName() {
   var random = Random().nextInt(pow(2, 32) as int).toString();
   return '${random}_tmp';
 }
 
+String? testTempPath;
 Future<Isar> openTempIsar(List<CollectionSchema<dynamic>> schemas,
     {String? name}) async {
-  registerBinaries();
+  await _prepareTest();
+  if (!kIsWeb && testTempPath == null) {
+    final dartToolDir = path.join(Directory.current.path, '.dart_tool');
+    testTempPath = path.join(dartToolDir, 'test', 'tmp');
+  }
 
-  return tOpen(
+  return await tOpen(
     schemas: schemas,
     name: name ?? getRandomName(),
     directory: kIsWeb ? '' : testTempPath!,